# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/channels:
    get:
      tags:
        - Channels
      summary: List Channels
      description: Get a paginated list of Channels with optional filtering and sorting
      operationId: listChannels
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Channels
      summary: Create Channel
      description: Create a new Channel
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/channel_members:
    get:
      tags:
        - ChannelMembers
      summary: List ChannelMembers
      description: Get a paginated list of ChannelMembers with optional filtering and sorting
      operationId: listChannelMembers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelMemberList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ChannelMembers
      summary: Create ChannelMember
      description: Create a new ChannelMember
      operationId: createChannelMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelMemberRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Channel:
      type: object
      required:
        - id
        - name
        - max_sessions_mode
        - max_sessions
        - block_assignment_during_call
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        website_id:
          nullable: true
          type: string
          format: uuid
        button_text:
          nullable: true
          type: string
          maxLength: 120
        welcome_message:
          nullable: true
          type: string
          maxLength: 1000
        max_sessions_mode:
          type: string
        max_sessions:
          type: integer
          format: int32
        block_assignment_during_call:
          type: boolean
        review_link:
          nullable: true
          type: string
          format: uri
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateChannelRequest:
      type: object
      required:
        - name
        - max_sessions_mode
        - max_sessions
        - block_assignment_during_call
        - is_active
        - metadata
      properties:
        name:
          type: string
          maxLength: 200
        website_id:
          nullable: true
          type: string
          format: uuid
        button_text:
          nullable: true
          type: string
          maxLength: 120
        welcome_message:
          nullable: true
          type: string
          maxLength: 1000
        max_sessions_mode:
          type: string
        max_sessions:
          type: integer
          format: int32
        block_assignment_during_call:
          type: boolean
        review_link:
          nullable: true
          type: string
          format: uri
        is_active:
          type: boolean
        metadata:
          type: object
...
    ChannelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ChannelMember:
      type: object
      required:
        - id
        - channel_id
        - user_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        channel_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateChannelMemberRequest:
      type: object
      required:
        - channel_id
        - user_id
        - metadata
      properties:
        channel_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    ChannelMemberList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChannelMember'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []