# src/presentation/http/channel_member_handler.rs
//! ChannelMember REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{ChannelMemberService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateChannelMemberDto, UpdateChannelMemberDto, PatchChannelMemberDto, ChannelMemberResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum ChannelMemberError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for ChannelMemberError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for ChannelMemberError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "CHANNELMEMBER_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "CHANNELMEMBER_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CHANNELMEMBER_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CHANNELMEMBER_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for ChannelMember.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /channel_members | List with pagination |
/// | POST | /channel_members | Create new |
/// | GET | /channel_members/:id | Get by ID |
/// | PUT | /channel_members/:id | Full update |
/// | PATCH | /channel_members/:id | Partial update |
/// | DELETE | /channel_members/:id | Soft delete |
/// | POST | /channel_members/bulk | Bulk create |
/// | POST | /channel_members/upsert | Upsert |
/// | GET | /channel_members/trash | List deleted |
/// | POST | /channel_members/:id/restore | Restore |
/// | DELETE | /channel_members/empty | Empty trash |
/// | GET | /channel_members/:id/deleted | Get deleted by ID |
/// | DELETE | /channel_members/trash/:id | Permanent delete from trash |
/// | GET | /channel_members/count | Count active entities |
/// | GET | /channel_members/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(ChannelMemberService::with_repository(repository));
/// let router = create_channel_member_routes(service);
/// ```
pub fn create_channel_member_routes(service: Arc<ChannelMemberService>) -> Router {
    BackboneCrudHandler::<ChannelMemberService, ChannelMember, CreateChannelMemberDto, UpdateChannelMemberDto, ChannelMemberResponseDto>::routes(
        service,
        "/channel_members",
    )
}

/// Create Axum router with only the read (GET) endpoints for ChannelMember.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_channel_member_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_channel_member_read_routes(service: Arc<ChannelMemberService>) -> Router {
    BackboneCrudHandler::<ChannelMemberService, ChannelMember, CreateChannelMemberDto, UpdateChannelMemberDto, ChannelMemberResponseDto>::read_routes(
        service,
        "/channel_members",
    )
}

/// Create Axum router with only the write (mutation) endpoints for ChannelMember.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_channel_member_write_routes(service: Arc<ChannelMemberService>) -> Router {
    BackboneCrudHandler::<ChannelMemberService, ChannelMember, CreateChannelMemberDto, UpdateChannelMemberDto, ChannelMemberResponseDto>::write_routes(
        service,
        "/channel_members",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_channel_member_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<ChannelMemberService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_channel_member_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}